feat(components): user-avatar
Added Dropdown menu and Avatar components
Created User avatar component - it's responsible for presenting the user profile picture.
The session status is now being fetched so we can provide personalized greetings and hints. Depend on that, Navbar will either show a sign-in button or a user menu with user's avatar.