feat(pages): intercept-signin
This commit involves several updates to improve user interface and interaction, and development workflow.

The main changes include;

Addition of a Dialog component to improve modals handling in the app. This is important for enhancing user interactions with the application.
Adding a check for user session on the HomePage. This ensures that user-specific information is displayed only when the browser has session data.
Modification of the SignIn link in the Navbar component to ensure it gets disabled when we are in the signin route.
Creation of a SignIn modal for user sign-in to apply a better visual user experience.
Clearing up Next.js build cache before each development and build command run. This is essential for eliminating potential stale cache issues during development or after deployment.